+++ .build.ps1
@@ -1,37 +1,19 @@
 
+param(
+	[string] $NuGetApiKey = $env:nuget_apikey,
+
+	# Overwrite published versions
+	[switch] $ForcePublish,
+
+    # Add doc templates for new command.
+	[switch] $ForceDocInit,
+
+	# Version suffix to prereleases
+	[int] $BuildNumber
+)
+
 $ModuleName = 'PsSqlLocalDb'
 
-. $PSScriptRoot/Tasks/Dependency.Tasks.ps1
-. $PSScriptRoot/Tasks/Build.Tasks.ps1
-
-task UpdateBuildTasks {
-    Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/Powershell/Build.Tasks.ps1' `
-        -OutFile "$PSScriptRoot\Tasks\Build.Tasks.ps1"
-}
-
-task UpdateVsCodeTasks {
-    Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/VsCode\tasks.json' `
-        -OutFile "$PSScriptRoot\.vscode\tasks.json"
-}
-
-task UpdateValidationWorkflow {
-    Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/build-validation-windows.yml' `
-        -OutFile "$PSScriptRoot\.github\workflows\build-validation.yml"
-}
-
-task UpdatePreReleaseWorkflow {
-    Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/pre-release-windows.yml' `
-        -OutFile "$PSScriptRoot\.github\workflows\pre-release.yml"
-}
-
-task UpdateReleaseWorkflow {
-    Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/release-windows.yml' `
-        -OutFile "$PSScriptRoot\.github\workflows\release.yml"
-}
-
-task UpdatePsBuildTasks -Jobs UpdateBuildTasks, UpdateVsCodeTasks, UpdateValidationWorkflow, UpdatePreReleaseWorkflow, UpdateReleaseWorkflow
+. $PSScriptRoot\tasks\Build.Tasks.ps1
+. $PSScriptRoot\tasks\Dependencies.Tasks.ps1
+. $PSScriptRoot\tasks\PsBuild.Tasks.ps1