+++ test/Test-Utility.Tests.ps1
@@ -1,12 +1,12 @@
 #Requires -Modules @{ ModuleName='Pester'; ModuleVersion='5.0.0' }
 
-Describe 'Test-Utility' {
+Describe Test-Utility {
 
     BeforeAll {
-        Import-Module $PSScriptRoot\..\Source\PsSqlLocalDb.psd1 -Force -ErrorAction Stop
+        Import-Module $PSScriptRoot\..\src\PsSqlLocalDb.psd1 -Force -ErrorAction Stop
     }
 
-    It 'works' {
+    It works {
         Test-LocalDbUtility | Should -BeIn $true, $false
     }